refactor(domain): split domain.ts into core + provider-specific type modules
- Extract core types (Social*, NormalizedStore, BackgroundStore, Export*) to shared/domain/core.ts
- Move X types (TweetData, XStore, ExportSummaryX) to providers/x/types.ts
- Move Instagram types (InstagramStore, ExportInstagramPost) to providers/instagram/types.ts
- Move LinkedIn types (LinkedInPostData, LinkedInStore, LinkedInExtra) to providers/linkedin/types.ts
- Create domain/index.ts barrel with re-exports for backward compat
- Replace domain.ts with re-export barrel

All 40+ existing imports from 'shared/domain' continue working via re-exports.
