+++ src/shared/domain/index.ts
@@ -1,0 +1,4 @@
+export * from "../../providers/instagram/types";
+export * from "../../providers/linkedin/types";
+export * from "../../providers/x/types";
+export * from "./core";
